re-orginized code for better debugging
