+++ InOut/Write.go
