--- InOut/write-here.go
@@ -1,37 +1,0 @@
-
-	package InOut
-	import "fmt"
-	
-		func SayHello0(){fmt.Printf("this is the salute 0")}
-		
-
-		func SayHello1(){fmt.Printf("this is the salute 1")}
-		
-
-		func SayHello2(){fmt.Printf("this is the salute 2")}
-		
-
-		func SayHello3(){fmt.Printf("this is the salute 3")}
-		
-
-		func SayHello4(){fmt.Printf("this is the salute 4")}
-		
-
-		func SayHello5(){fmt.Printf("this is the salute 5")}
-		
-
-		func SayHello6(){fmt.Printf("this is the salute 6")}
-		
-
-		func SayHello7(){fmt.Printf("this is the salute 7")}
-		
-
-		func SayHello8(){fmt.Printf("this is the salute 8")}
-		
-
-		func SayHello9(){fmt.Printf("this is the salute 9")}
-		
-
-		func SayHello10(){fmt.Printf("this is the salute 10")}
-		
-var Fs []CreatedFunc = []CreatedFunc {SayHello10,SayHello9,SayHello8,SayHello7,SayHello6,SayHello5,SayHello4,SayHello3,SayHello2,SayHello1,SayHello0,}
