+++ main.go
@@ -116,13 +116,6 @@ func main() {
 	InOut.OsWriter("/workspaces/golang-learning/InOut/write-here.txt", stringSlicer[0:10])
 	InOut.ReadFile("/tmp/numbers.txt")
 	InOut.OsOffSet("/tmp/numbers.txt")
-	var writeRecieve chan map[string]string = make(chan map[string]string, 2)
-	go InOut.RecieveAndWrite(
-		writeRecieve,
-		"/workspaces/golang-learning/InOut/write-here.go",
-		waiter,
-	)
-	go InOut.SendMessage(writeRecieve, waiter)
 	waiter.Wait()
 	fmt.Printf("final count %v\n", integer)
 	fmt.Printf("%v\n", *stringSlice)
@@ -135,10 +128,7 @@ func main() {
 	fmt.Printf("%v\n", e)
 	var stringToBeManipulated string = "faBiAnJeEsus"
 	stringManipulation.StringToUpperCase(stringToBeManipulated)
-	for _, funcMember := range InOut.Fs {
-		funcMember()
-	}
-	InOut.Write()
+	InOut.WriterStd()
 	InOut.ReaderStd()
 	InOut.BufferReader()
 	InOut.Scanner()
